Some more renamings and updated to latest compatibility date 2025-09-30

+++ src/main/java/net/troja/eve/esi/api/SsoApi.java
@@ -50,7 +50,7 @@ public class SsoApi {
         return evesso.getJWT().getPayload().getName();
     }
 
-    public Integer getCharacterId() {
+    public Long getCharacterId() {
         OAuth evesso = (OAuth) apiClient.getAuthentication(EVESSO);
         return evesso.getJWT().getPayload().getCharacterID();
     }
